check/requirements.txt
# pip install -r requirements.txt
lanzou-api==2.6.7.1
